Fix saveless saving and quitting hiding buttons
Basically, when you save and quit, if you have a save, it attempts to
return to the `play` menu. Unfortunately for us, if you don't have a
save first, that menu never existed. You went directly from `mainmenu`
to the gameplay state. Since `play` never existed, the code didn't
have any menu to return to, so no buttons got created.

My fix is simple: if the menu isn't found, just make it.