Glitchless: Disable R key
This disables pressing R to instantly die on the spot in glitchless
mode. However, it is still available in custom levels.

The R key was intended to get players out of softlocks in custom levels,
but it was left enabled in the main game by mistake. This turned out to
enable a variety of various glitches, most notably telejumping. So it
was left enabled.

But now glitchless mode disables it, to prevent pressing it by accident
for speedrunning goals that prohibit pressing R.