Glitchless: Fix warp token death warps
The warp token death warp glitch is simple. It takes advantage of the
fact that, in the main game, warp tokens (barring some exceptions) only
change the room but do not change the player's position. So if the
player dies while touching a warp token, the room will change but they
will stay at the same position of their respawn point, which is used in
speedrunning.

To fix this, if the player dies, make sure that the warp tokens in the
room don't warp them.

Otherwise, this would still be possible to do even without pressing R.