Glitchless: Fix line clipping
The reason that line clipping works is because when the gravity line
unflips the player, they are still eligible to flip (with
onground/onroof) so by doing it frame-perfectly they can flip again.

To fix this, just zero out their onground/onroof.

This glitch mainly applies to horizontal gravity lines but it is also
possible, but less severe (in most cases), with vertical gravity lines,
so this fixes it for vertical lines too.